Move intent parsing to TermuxService class
